dom/tests/mochitest/mouselock/test_movement.html

<!DOCTYPE HTML>
<html>
  <!-- https://bugzilla.mozilla.org/show_bug.cgi?id=92264 -->
    <head>
    <title>
      Test for Bug 92264
    </title>
    <script src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script src="/tests/SimpleTest/EventUtils.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  
  <body onload="runTests();">
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <p id="display">
    </p>
    <div id="content" style="display: none">
    </div>
    <canvas id="canvas" width="150" height="150">
    </canvas>
    <pre id="test">
      <script type="application/javascript">
			  /** Test for Bug 633602 **/
			  /** This is a test for correct values in the movementX and movementY
				    attributes in the mousemove event **/
        SimpleTest.waitForExplicitFinish();
        var pointer = navigator.pointer;
        var canvas = document.getElementById("canvas");
        var testMouseMovement = false;
        var mouseX = 0;
        var mouseY = 0;
        var centerX = parseInt(window.screen.width / 2);
        var centerY = parseInt(window.screen.height / 2);
        var tests = 0;
        var init = false;
        SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
          false);
        document.addEventListener("mousemove", function (e) {
          if (testMouseMovement === true) {
            if (tests === 0) {
              tests++;
              ok("movementX" in e && "movementY" in e,
                "movementX and movementY should exist in mouse events.");
              synthesizeMouse(canvas, centerX + 150, centerY - 50, {
                type: "mousemove"
              }, window);
            } else {
              if (tests === 1) {
                tests++;
                ok(e.movementX === 150 && e.movementY === -50,
                  "Movement values should equal to eNow - ePrevious.");
                pointer.unlock();
                mouseX = e.screenX;
                mouseY = e.screenY;
                synthesizeMouse(canvas, mouseX + 200, mouseY + 50, {
                  type: "mousemove"
                }, window);
              } else if (tests === 2) {
                tests++;
                ok(e.movementX === 200 && e.movementY === 50, 
                  "Movement values should equal to eNow - ePrevious.");
                document.mozCancelFullScreen();
              } else if (tests > 2) {
                document.mozCancelFullScreen();
              }
            }
          }
        });
        document.addEventListener("mozfullscreenchange", function(e) {
          if (document.mozFullScreen) {
            if (document.mozFullScreenElement === canvas) {
              testMouseMovement = true;
              pointer.lock(canvas);
              synthesizeMouse(canvas, 0, 0, {
                type: "mousemove"
              }, window);
            }
          } else {
            SimpleTest.finish();
          }
        }, false);
        function runTests() {
          canvas.mozRequestFullScreen();
        }
      </script>
    </pre>
  </body>
</html>
